feat: Create states and behaviors within classes Board and Square
Class Board: Create a state @squares within initialize method and set it to a hash with key, value pairs
Keys are numbered from 1 to 9 and values are set to a class Square object. Create a method called get_square_at(key) that takes a key value as an argument and returns the value stored for that particular key.

Class Square: Create a state called marker. The value for the state marker is provided by the class Board's instance variables squares.